Add support for OSX keychain saving and loading of your password.
Using the OSX keychain gem from https://github.com/seattlerb/osx_keychain
just requests the password from keychain and puts it back there if it
works.

Also stores the username and service in the YAML, so you could read it out
of there when you have to reauth, and have no user input. But I didn't
write that.